cherry-pick or reverted #2581 #2590 #2600 and #2603
