+++ paasta_tools/kubernetes_tools.py
@@ -128,6 +128,7 @@ log = logging.getLogger(__name__)
 
 KUBE_CONFIG_PATH = "/etc/kubernetes/admin.conf"
 YELP_ATTRIBUTE_PREFIX = "yelp.com/"
+PAASTA_ATTRIBUTE_PREFIX = "paasta.yelp.com/"
 CONFIG_HASH_BLACKLIST = {"replicas"}
 KUBE_DEPLOY_STATEGY_MAP = {
     "crossover": "RollingUpdate",
@@ -136,6 +137,7 @@ KUBE_DEPLOY_STATEGY_MAP = {
 }
 HACHECK_POD_NAME = "hacheck"
 KUBERNETES_NAMESPACE = "paasta"
+DISCOVERY_ATTRIBUTES = {"region", "superregion", "ecosystem", "habitat"}
 
 
 # For detail, https://github.com/kubernetes-client/python/issues/553
@@ -1589,11 +1591,17 @@ def get_pod_status(pod: V1Pod,) -> PodStatus:
     return _POD_STATUS_NAME_TO_STATUS[pod.status.phase.upper()]
 
 
-def get_active_shas_for_service(pod_list: Sequence[V1Pod],) -> Mapping[str, Set[str]]:
-    ret: Mapping[str, Set[str]] = {"config_sha": set(), "git_sha": set()}
-    for pod in pod_list:
-        ret["config_sha"].add(pod.metadata.labels["paasta.yelp.com/config_sha"])
-        ret["git_sha"].add(pod.metadata.labels["paasta.yelp.com/git_sha"])
+def get_active_shas_for_service(
+    obj_list: Sequence[Union[V1Pod, V1ReplicaSet, V1Deployment, V1StatefulSet]],
+) -> Mapping[str, Set[str]]:
+    ret: MutableMapping[str, Set[str]] = {"config_sha": set(), "git_sha": set()}
+    for obj in obj_list:
+        config_sha = obj.metadata.labels.get("paasta.yelp.com/config_sha")
+        if config_sha is not None:
+            ret["config_sha"].add(config_sha)
+        git_sha = obj.metadata.labels.get("paasta.yelp.com/git_sha")
+        if git_sha is not None:
+            ret["git_sha"].add(git_sha)
     return ret
 
 
@@ -1614,8 +1622,8 @@ def filter_nodes_by_blacklist(
     :returns: The list of nodes after the filter
     """
     if whitelist:
-        whitelist = (maybe_add_yelp_prefix(whitelist[0]), whitelist[1])
-    blacklist = [(maybe_add_yelp_prefix(entry[0]), entry[1]) for entry in blacklist]
+        whitelist = (paasta_prefixed(whitelist[0]), whitelist[1])
+    blacklist = [(paasta_prefixed(entry[0]), entry[1]) for entry in blacklist]
     return [
         node
         for node in nodes
@@ -1624,14 +1632,20 @@ def filter_nodes_by_blacklist(
     ]
 
 
-def maybe_add_yelp_prefix(attribute: str,) -> str:
-    return YELP_ATTRIBUTE_PREFIX + attribute if "/" not in attribute else attribute
+def paasta_prefixed(attribute: str,) -> str:
+    # discovery attributes are exempt for now
+    if attribute in DISCOVERY_ATTRIBUTES:
+        return YELP_ATTRIBUTE_PREFIX + attribute
+    elif "/" in attribute:
+        return attribute
+    else:
+        return PAASTA_ATTRIBUTE_PREFIX + attribute
 
 
 def get_nodes_grouped_by_attribute(
     nodes: Sequence[V1Node], attribute: str
 ) -> Mapping[str, Sequence[V1Node]]:
-    attribute = maybe_add_yelp_prefix(attribute)
+    attribute = paasta_prefixed(attribute)
     sorted_nodes = sorted(
         nodes, key=lambda node: node.metadata.labels.get(attribute, "")
     )
@@ -1885,7 +1899,9 @@ def sanitised_cr_name(service: str, instance: str) -> str:
     return f"{sanitised_service}-{sanitised_instance}"
 
 
-def get_cr(kube_client: KubeClient, cr_id: dict) -> Optional[Mapping[str, Any]]:
+def get_cr(
+    kube_client: KubeClient, cr_id: Mapping[str, str]
+) -> Optional[Mapping[str, Any]]:
     try:
         return kube_client.custom.get_namespaced_custom_object(**cr_id)
     except ApiException as e:
@@ -1895,18 +1911,20 @@ def get_cr(kube_client: KubeClient, cr_id: dict) -> Optional[Mapping[str, Any]]:
             raise
 
 
-def get_cr_status(kube_client: KubeClient, cr_id: dict) -> Optional[Mapping[str, Any]]:
+def get_cr_status(
+    kube_client: KubeClient, cr_id: Mapping[str, str]
+) -> Optional[Mapping[str, Any]]:
     return (get_cr(kube_client, cr_id) or {}).get("status")
 
 
 def get_cr_metadata(
-    kube_client: KubeClient, cr_id: dict
+    kube_client: KubeClient, cr_id: Mapping[str, str]
 ) -> Optional[Mapping[str, Any]]:
     return (get_cr(kube_client, cr_id) or {}).get("metadata")
 
 
 def set_cr_desired_state(
-    kube_client: KubeClient, cr_id: dict, desired_state: str
+    kube_client: KubeClient, cr_id: Mapping[str, str], desired_state: str
 ) -> str:
     cr = kube_client.custom.get_namespaced_custom_object(**cr_id)
     if cr.get("status", {}).get("state") == desired_state:
